[BottomSheet] Specifying a trackingScrollView is now correctly passed to the Presentation Controller.
The presentation controller is created on the first call to mdc_bottomSheetPresentationController. This is (as of #3325) being called from setDismissOnBackgroundTap, which when using a Bottom Sheet Controller is called during initialization. This results in the Bottom Sheet controller's trackingScrollView property to be ignored and instead relies on the discovery of a scroll view based on the contentViewController type.